chore: refactored components, removed unused imports, renamed folders, moved media files to assets folder
- Refactored components for improved code readability and maintainability
- Removed unused imports to clean up the codebase
- Renamed folders to follow the correct naming convention
- Moved media files to the assets folder for better organization
