--- RediUX-Front/src/components/10ResultadosPesquisa/folder.png
@@ -1,4 +1,0 @@
-<svg width="240" height="120" viewBox="0 0 240 120" fill="none" xmlns="http://www.w3.org/2000/svg">
-<rect width="240" height="120" fill="#BECBEA"/>
-<path fill-rule="evenodd" clip-rule="evenodd" d="M97.5 37.5C97.5 35.5109 98.2902 33.6032 99.6967 32.1967C101.103 30.7902 103.011 30 105 30H124.849C125.843 30.0002 126.797 30.3954 127.5 31.0988L141.401 45C142.105 45.7031 142.5 46.6568 142.5 47.6512V82.5C142.5 84.4891 141.71 86.3968 140.303 87.8033C138.897 89.2098 136.989 90 135 90H105C103.011 90 101.103 89.2098 99.6967 87.8033C98.2902 86.3968 97.5 84.4891 97.5 82.5V37.5ZM123.75 45L125.625 35.625L136.875 46.875L127.5 48.75C126.505 48.75 125.552 48.3549 124.848 47.6516C124.145 46.9484 123.75 45.9946 123.75 45ZM114.375 50.625C114.375 50.1277 114.177 49.6508 113.826 49.2992C113.474 48.9475 112.997 48.75 112.5 48.75C112.003 48.75 111.526 48.9475 111.174 49.2992C110.823 49.6508 110.625 50.1277 110.625 50.625V53.0025L108.566 51.8137C108.353 51.6887 108.117 51.6071 107.872 51.5736C107.627 51.5401 107.378 51.5554 107.139 51.6186C106.9 51.6818 106.675 51.7916 106.479 51.9418C106.283 52.092 106.118 52.2796 105.994 52.4938C105.871 52.7079 105.79 52.9444 105.759 53.1896C105.727 53.4348 105.744 53.6839 105.808 53.9225C105.873 54.1612 105.985 54.3846 106.136 54.5801C106.288 54.7755 106.476 54.939 106.691 55.0613L108.75 56.25L106.691 57.4387C106.476 57.561 106.288 57.7245 106.136 57.9199C105.985 58.1154 105.873 58.3388 105.808 58.5775C105.744 58.8161 105.727 59.0652 105.759 59.3104C105.79 59.5556 105.871 59.7921 105.994 60.0062C106.118 60.2204 106.283 60.408 106.479 60.5582C106.675 60.7084 106.9 60.8182 107.139 60.8814C107.378 60.9446 107.627 60.9599 107.872 60.9264C108.117 60.8929 108.353 60.8113 108.566 60.6862L110.625 59.4975V61.875C110.625 62.3723 110.823 62.8492 111.174 63.2008C111.526 63.5525 112.003 63.75 112.5 63.75C112.997 63.75 113.474 63.5525 113.826 63.2008C114.177 62.8492 114.375 62.3723 114.375 61.875V59.4975L116.434 60.6862C116.647 60.8113 116.883 60.8929 117.128 60.9264C117.373 60.9599 117.622 60.9446 117.861 60.8814C118.1 60.8182 118.325 60.7084 118.521 60.5582C118.717 60.408 118.882 60.2204 119.006 60.0062C119.129 59.7921 119.21 59.5556 119.241 59.3104C119.273 59.0652 119.256 58.8161 119.192 58.5775C119.127 58.3388 119.015 58.1154 118.864 57.9199C118.712 57.7245 118.524 57.561 118.309 57.4387L116.25 56.25L118.309 55.0613C118.524 54.939 118.712 54.7755 118.864 54.5801C119.015 54.3846 119.127 54.1612 119.192 53.9225C119.256 53.6839 119.273 53.4348 119.241 53.1896C119.21 52.9444 119.129 52.7079 119.006 52.4938C118.882 52.2796 118.717 52.092 118.521 51.9418C118.325 51.7916 118.1 51.6818 117.861 51.6186C117.622 51.5554 117.373 51.5401 117.128 51.5736C116.883 51.6071 116.647 51.6887 116.434 51.8137L114.375 53.0025V50.625ZM106.875 67.5C106.378 67.5 105.901 67.6975 105.549 68.0492C105.198 68.4008 105 68.8777 105 69.375C105 69.8723 105.198 70.3492 105.549 70.7008C105.901 71.0525 106.378 71.25 106.875 71.25H125.625C126.122 71.25 126.599 71.0525 126.951 70.7008C127.302 70.3492 127.5 69.8723 127.5 69.375C127.5 68.8777 127.302 68.4008 126.951 68.0492C126.599 67.6975 126.122 67.5 125.625 67.5H106.875ZM106.875 75C106.378 75 105.901 75.1975 105.549 75.5492C105.198 75.9008 105 76.3777 105 76.875C105 77.3723 105.198 77.8492 105.549 78.2008C105.901 78.5525 106.378 78.75 106.875 78.75H125.625C126.122 78.75 126.599 78.5525 126.951 78.2008C127.302 77.8492 127.5 77.3723 127.5 76.875C127.5 76.3777 127.302 75.9008 126.951 75.5492C126.599 75.1975 126.122 75 125.625 75H106.875Z" fill="#6489D1"/>
-</svg>
