+++ RediUX-Front/src/components/ConteudoADM/ConteudoADM.jsx
@@ -1,8 +1,8 @@
 import { Box, Chip, Container, Button, Typography, Stack } from "@mui/material";
 import { InsertDriveFileOutlined, InsertLink, List, PersonOutline, ArrowBackIosNew } from "@mui/icons-material";
-import folder from "./folder.svg"
+import folder from "../../assets/folder.svg"
 import { Link } from "react-router-dom";
-import Tollbaradm from "../00TollbarADM/TollbarADM";
+import Tollbaradm from "../TollbarADM/TollbarADM";
 import { useEffect } from "react"
 import { useState } from "react"
 import { useParams } from "react-router-dom"
