+++ RediUX-Front/src/components/ListaConteudos/ListaConteudos.jsx
@@ -1,13 +1,15 @@
-import { Button, Box, Container, InputAdornment, TextField, FormControl, InputLabel, Select, MenuItem, Typography, Card, CardMedia, CardContent, CardActions, Chip, IconButton } from "@mui/material";
+import { Button, Box, Container, InputAdornment, TextField, FormControl, InputLabel, Select, MenuItem, Typography, Card, CardMedia, CardContent, CardActions, Chip, Divider, IconButton } from "@mui/material";
 import SearchIcon from "@mui/icons-material/Search";
 import { useState } from "react";
-import folder from "./folder.png"
+import folder from "../../assets/folder.svg";
 import { ArrowBackIos } from "@mui/icons-material";
 import { Link } from "react-router-dom";
-import Tollbar from "../00ToolbarU/Toolbar";
+import Tollbar from "../ToolbarU/Toolbar";
+import { useNavigate } from "react-router-dom";
 
-const Artigo = () => {
+const ListaConteudos = () => {
     const [search, setSearch] = useState("");
+    const navigate = useNavigate();
 
     const handleChange = (event) => {
         setSearch(event.target.value);
@@ -19,11 +21,36 @@ const Artigo = () => {
         setTag(event.target.value);
     };
 
+    const handleKeyPress = (event) => {
+        if (event.key === "Enter") {
+            handleSubmit(event);
+        }
+    };
+
+    const handleSubmit = (event) => {
+        event.preventDefault();
+        const searchTerm = search.trim();
+        navigate(`/results?term=${searchTerm}`);
+    };
+
     return (
         <>
             <Tollbar />
 
             <Container>
+                <Typography
+                    component="h1"
+                    variant="h4"
+                    fontWeight={500}
+                    mt={10}
+                    sx={{
+                        color: "#8B8B8"
+                    }}
+                >
+                    Mostrando resultados para:
+                    "Pesquisa do Usuário"
+                </Typography>
+
                 <Box
                     sx={{
                         display: "flex",
@@ -31,6 +58,7 @@ const Artigo = () => {
                         mt: 5,
                     }}
                 >
+
                     <Link to="/">
                         <IconButton
                             sx={{
@@ -49,12 +77,14 @@ const Artigo = () => {
                         label="Pesquisar Conteúdo"
                         value={search}
                         onChange={handleChange}
+                        onKeyDown={handleKeyPress}
                         sx={{ width: 860 }}
                         InputProps={{
                             endAdornment: (
                                 <InputAdornment position="end">
                                     <Button
                                         sx={{ height: 55 }}
+                                        onClick={handleSubmit}
                                     >
                                         <SearchIcon />
                                     </Button>
@@ -96,14 +126,13 @@ const Artigo = () => {
                             mt: 5,
                             mb: 2,
                         }}>
-                        Artigos
+                        Livros
                     </Typography>
                     <Box
                         sx={{
                             display: "flex",
                             overflow: "auto",
-                            mb: 5,
-                            pb: 2,
+                            mb: 5
                         }}
                     >
                         <Card sx={{ minWidth: 240, mr: 2 }}>
@@ -120,21 +149,45 @@ const Artigo = () => {
                                 <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
                             </CardContent>
                             <CardActions>
-                                <Button
-                                    size="small"
-                                    sx={{
-                                        backgroundColor: "#0C2D8A",
-                                        color: "#BECBEA",
-                                        '&:hover': {
-                                            backgroundColor: "#BECBEA",
-                                            color: "#0C2D8A",
-                                        }
-                                    }}
-                                >
-                                    Visualizar
-                                </Button>
+                                <Link to="/IDConteudo">
+                                    <Button
+                                        size="small"
+                                        sx={{
+                                            backgroundColor: "#0C2D8A",
+                                            color: "#BECBEA",
+                                            '&:hover': {
+                                                backgroundColor: "#BECBEA",
+                                                color: "#0C2D8A",
+                                            }
+                                        }}
+                                    >
+                                        Visualizar
+                                    </Button>
+                                </Link>
                             </CardActions>
                         </Card>
+                    </Box>
+                </Container>
+
+                <Divider />
+
+                <Container>
+                    <Typography
+                        variant="h5"
+                        component="h2"
+                        sx={{
+                            mt: 5,
+                            mb: 2,
+                        }}>
+                        Artigos
+                    </Typography>
+                    <Box
+                        sx={{
+                            display: "flex",
+                            overflow: "auto",
+                            mb: 5,
+                        }}
+                    >
                         <Card sx={{ minWidth: 240, mr: 2 }}>
                             <CardMedia
                                 sx={{ height: 140 }}
@@ -145,7 +198,7 @@ const Artigo = () => {
                                 <Typography gutterBottom variant="h6" component="div">
                                     Título
                                 </Typography>
-                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1, }} />
+                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1 }} />
                                 <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
                             </CardContent>
                             <CardActions>
@@ -164,6 +217,28 @@ const Artigo = () => {
                                 </Button>
                             </CardActions>
                         </Card>
+                    </Box>
+                </Container>
+
+                <Divider />
+
+                <Container>
+                    <Typography
+                        variant="h5"
+                        component="h2"
+                        sx={{
+                            mt: 5,
+                            mb: 2,
+                        }}>
+                        Podcasts
+                    </Typography>
+                    <Box
+                        sx={{
+                            display: "flex",
+                            overflow: "auto",
+                            mb: 5,
+                        }}
+                    >
                         <Card sx={{ minWidth: 240, mr: 2 }}>
                             <CardMedia
                                 sx={{ height: 140 }}
@@ -174,7 +249,7 @@ const Artigo = () => {
                                 <Typography gutterBottom variant="h6" component="div">
                                     Título
                                 </Typography>
-                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1, }} />
+                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1 }} />
                                 <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
                             </CardContent>
                             <CardActions>
@@ -193,6 +268,28 @@ const Artigo = () => {
                                 </Button>
                             </CardActions>
                         </Card>
+                    </Box>
+                </Container>
+
+                <Divider />
+
+                <Container>
+                    <Typography
+                        variant="h5"
+                        component="h2"
+                        sx={{
+                            mt: 5,
+                            mb: 2,
+                        }}>
+                        Vídeos
+                    </Typography>
+                    <Box
+                        sx={{
+                            display: "flex",
+                            overflow: "auto",
+                            mb: 5,
+                        }}
+                    >
                         <Card sx={{ minWidth: 240, mr: 2 }}>
                             <CardMedia
                                 sx={{ height: 140 }}
@@ -203,36 +300,7 @@ const Artigo = () => {
                                 <Typography gutterBottom variant="h6" component="div">
                                     Título
                                 </Typography>
-                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1, }} />
-                                <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
-                            </CardContent>
-                            <CardActions>
-                                <Button
-                                    size="small"
-                                    sx={{
-                                        backgroundColor: "#0C2D8A",
-                                        color: "#BECBEA",
-                                        '&:hover': {
-                                            backgroundColor: "#BECBEA",
-                                            color: "#0C2D8A",
-                                        }
-                                    }}
-                                >
-                                    Visualizar
-                                </Button>
-                            </CardActions>
-                        </Card>
-                        <Card sx={{ minWidth: 240, mr: 2 }}>
-                            <CardMedia
-                                sx={{ height: 140 }}
-                                image={folder}
-                                tittle="Conteúdo"
-                            />
-                            <CardContent>
-                                <Typography gutterBottom variant="h6" component="div">
-                                    Título
-                                </Typography>
-                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1, }} />
+                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1 }} />
                                 <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
                             </CardContent>
                             <CardActions>
@@ -258,4 +326,4 @@ const Artigo = () => {
     )
 }
 
-export default Artigo
+export default ListaConteudos
