+++ RediUX-Front/src/components/PagMidias/Video.jsx
@@ -1,15 +1,13 @@
-import { Button, Box, Container, InputAdornment, TextField, FormControl, InputLabel, Select, MenuItem, Typography, Card, CardMedia, CardContent, CardActions, Chip, Divider, IconButton } from "@mui/material";
+import { Button, Box, Container, InputAdornment, TextField, FormControl, InputLabel, Select, MenuItem, Typography, Card, CardMedia, CardContent, CardActions, Chip, IconButton } from "@mui/material";
 import SearchIcon from "@mui/icons-material/Search";
 import { useState } from "react";
-import folder from "./folder.png"
+import folder from "../../assets/folder.svg";
 import { ArrowBackIos } from "@mui/icons-material";
 import { Link } from "react-router-dom";
-import Tollbar from "../00ToolbarU/Toolbar";
-import { useNavigate } from "react-router-dom";
+import Tollbar from "../ToolbarU/Toolbar";
 
-const ListaConteudos = () => {
+const Video = () => {
     const [search, setSearch] = useState("");
-    const navigate = useNavigate();
 
     const handleChange = (event) => {
         setSearch(event.target.value);
@@ -21,36 +19,11 @@ const ListaConteudos = () => {
         setTag(event.target.value);
     };
 
-    const handleKeyPress = (event) => {
-        if (event.key === "Enter") {
-            handleSubmit(event);
-        }
-    };
-
-    const handleSubmit = (event) => {
-        event.preventDefault();
-        const searchTerm = search.trim();
-        navigate(`/results?term=${searchTerm}`);
-    };
-
     return (
         <>
             <Tollbar />
-
+            
             <Container>
-                <Typography
-                    component="h1"
-                    variant="h4"
-                    fontWeight={500}
-                    mt={10}
-                    sx={{
-                        color: "#8B8B8"
-                    }}
-                >
-                    Mostrando resultados para:
-                    "Pesquisa do Usuário"
-                </Typography>
-
                 <Box
                     sx={{
                         display: "flex",
@@ -58,7 +31,6 @@ const ListaConteudos = () => {
                         mt: 5,
                     }}
                 >
-
                     <Link to="/">
                         <IconButton
                             sx={{
@@ -77,14 +49,12 @@ const ListaConteudos = () => {
                         label="Pesquisar Conteúdo"
                         value={search}
                         onChange={handleChange}
-                        onKeyDown={handleKeyPress}
                         sx={{ width: 860 }}
                         InputProps={{
                             endAdornment: (
                                 <InputAdornment position="end">
                                     <Button
                                         sx={{ height: 55 }}
-                                        onClick={handleSubmit}
                                     >
                                         <SearchIcon />
                                     </Button>
@@ -126,13 +96,14 @@ const ListaConteudos = () => {
                             mt: 5,
                             mb: 2,
                         }}>
-                        Livros
+                        Vídeos
                     </Typography>
                     <Box
                         sx={{
                             display: "flex",
                             overflow: "auto",
-                            mb: 5
+                            mb: 5,
+                            pb: 2,
                         }}
                     >
                         <Card sx={{ minWidth: 240, mr: 2 }}>
@@ -149,45 +120,21 @@ const ListaConteudos = () => {
                                 <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
                             </CardContent>
                             <CardActions>
-                                <Link to="/IDConteudo">
-                                    <Button
-                                        size="small"
-                                        sx={{
-                                            backgroundColor: "#0C2D8A",
-                                            color: "#BECBEA",
-                                            '&:hover': {
-                                                backgroundColor: "#BECBEA",
-                                                color: "#0C2D8A",
-                                            }
-                                        }}
-                                    >
-                                        Visualizar
-                                    </Button>
-                                </Link>
+                                <Button
+                                    size="small"
+                                    sx={{
+                                        backgroundColor: "#0C2D8A",
+                                        color: "#BECBEA",
+                                        '&:hover': {
+                                            backgroundColor: "#BECBEA",
+                                            color: "#0C2D8A",
+                                        }
+                                    }}
+                                >
+                                    Visualizar
+                                </Button>
                             </CardActions>
                         </Card>
-                    </Box>
-                </Container>
-
-                <Divider />
-
-                <Container>
-                    <Typography
-                        variant="h5"
-                        component="h2"
-                        sx={{
-                            mt: 5,
-                            mb: 2,
-                        }}>
-                        Artigos
-                    </Typography>
-                    <Box
-                        sx={{
-                            display: "flex",
-                            overflow: "auto",
-                            mb: 5,
-                        }}
-                    >
                         <Card sx={{ minWidth: 240, mr: 2 }}>
                             <CardMedia
                                 sx={{ height: 140 }}
@@ -198,7 +145,7 @@ const ListaConteudos = () => {
                                 <Typography gutterBottom variant="h6" component="div">
                                     Título
                                 </Typography>
-                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1 }} />
+                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1, }} />
                                 <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
                             </CardContent>
                             <CardActions>
@@ -217,28 +164,6 @@ const ListaConteudos = () => {
                                 </Button>
                             </CardActions>
                         </Card>
-                    </Box>
-                </Container>
-
-                <Divider />
-
-                <Container>
-                    <Typography
-                        variant="h5"
-                        component="h2"
-                        sx={{
-                            mt: 5,
-                            mb: 2,
-                        }}>
-                        Podcasts
-                    </Typography>
-                    <Box
-                        sx={{
-                            display: "flex",
-                            overflow: "auto",
-                            mb: 5,
-                        }}
-                    >
                         <Card sx={{ minWidth: 240, mr: 2 }}>
                             <CardMedia
                                 sx={{ height: 140 }}
@@ -249,7 +174,7 @@ const ListaConteudos = () => {
                                 <Typography gutterBottom variant="h6" component="div">
                                     Título
                                 </Typography>
-                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1 }} />
+                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1, }} />
                                 <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
                             </CardContent>
                             <CardActions>
@@ -268,28 +193,6 @@ const ListaConteudos = () => {
                                 </Button>
                             </CardActions>
                         </Card>
-                    </Box>
-                </Container>
-
-                <Divider />
-
-                <Container>
-                    <Typography
-                        variant="h5"
-                        component="h2"
-                        sx={{
-                            mt: 5,
-                            mb: 2,
-                        }}>
-                        Vídeos
-                    </Typography>
-                    <Box
-                        sx={{
-                            display: "flex",
-                            overflow: "auto",
-                            mb: 5,
-                        }}
-                    >
                         <Card sx={{ minWidth: 240, mr: 2 }}>
                             <CardMedia
                                 sx={{ height: 140 }}
@@ -300,7 +203,36 @@ const ListaConteudos = () => {
                                 <Typography gutterBottom variant="h6" component="div">
                                     Título
                                 </Typography>
-                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1 }} />
+                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1, }} />
+                                <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
+                            </CardContent>
+                            <CardActions>
+                                <Button
+                                    size="small"
+                                    sx={{
+                                        backgroundColor: "#0C2D8A",
+                                        color: "#BECBEA",
+                                        '&:hover': {
+                                            backgroundColor: "#BECBEA",
+                                            color: "#0C2D8A",
+                                        }
+                                    }}
+                                >
+                                    Visualizar
+                                </Button>
+                            </CardActions>
+                        </Card>
+                        <Card sx={{ minWidth: 240, mr: 2 }}>
+                            <CardMedia
+                                sx={{ height: 140 }}
+                                image={folder}
+                                tittle="Conteúdo"
+                            />
+                            <CardContent>
+                                <Typography gutterBottom variant="h6" component="div">
+                                    Título
+                                </Typography>
+                                <Chip label="Tags" sx={{ backgroundColor: "#BECBEA", color: "0C2D8A", mr: 1, }} />
                                 <Chip label="Tags" sx={{ backgroundColor: "#FFB1B1", color: "#BF4040", mr: 1 }} />
                             </CardContent>
                             <CardActions>
@@ -326,4 +258,4 @@ const ListaConteudos = () => {
     )
 }
 
-export default ListaConteudos
+export default Video
